+++ atlasdb-client/src/main/java/com/palantir/atlasdb/keyvalue/impl/DualWriteKeyValueService.java
@@ -278,4 +278,15 @@ public class DualWriteKeyValueService implements KeyValueService {
                                                      long timestamp) {
         return delegate1.getRowsColumnRange(tableRef, rows, columnRangeSelection, cellBatchHint, timestamp);
     }
+
+    @Override
+    public void compactInternally(TableReference tableRef, boolean inMaintenanceMode) {
+        delegate1.compactInternally(tableRef, inMaintenanceMode);
+        delegate2.compactInternally(tableRef, inMaintenanceMode);
+    }
+
+    @Override
+    public boolean shouldTriggerCompactions() {
+        return delegate1.shouldTriggerCompactions() || delegate2.shouldTriggerCompactions();
+    }
 }