+++ atlasdb-client/src/main/java/com/palantir/atlasdb/keyvalue/impl/TracingKeyValueService.java
@@ -116,6 +116,13 @@ public final class TracingKeyValueService extends ForwardingObject implements Ke
         }
     }
 
+    public void compactInternally(TableReference tableRef, boolean inMaintenanceMode) {
+        //noinspection unused - try-with-resources closes trace
+        try (CloseableTrace trace = startLocalTrace("compactInternally({})", tableRef)) {
+            delegate().compactInternally(tableRef, inMaintenanceMode);
+        }
+    }
+
     @Override
     public boolean isInitialized() {
         return delegate().isInitialized();
@@ -391,6 +398,11 @@ public final class TracingKeyValueService extends ForwardingObject implements Ke
                 LoggingArgs.safeTablesOrPlaceholder(tableRefs))) {
             delegate().truncateTables(tableRefs);
         }
+    }
+
+    @Override
+    public boolean shouldTriggerCompactions() {
+        return delegate().shouldTriggerCompactions();
     }
 }