+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/compact/BackgroundCompactor.java
@@ -16,75 +16,103 @@
 
 package com.palantir.atlasdb.compact;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ThreadLocalRandom;
+import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.logging.LoggingArgs;
 import com.palantir.atlasdb.schema.generated.CompactMetadataTable;
 import com.palantir.atlasdb.schema.generated.CompactTableFactory;
-import com.palantir.atlasdb.schema.generated.SweepPriorityTable;
-import com.palantir.atlasdb.schema.generated.SweepTableFactory;
-import com.palantir.atlasdb.transaction.api.Transaction;
 import com.palantir.atlasdb.transaction.api.TransactionManager;
+import com.palantir.common.base.Throwables;
 import com.palantir.lock.LockService;
 import com.palantir.lock.SingleLockService;
+import com.palantir.logsafe.SafeArg;
 
-public class BackgroundCompactor implements Runnable {
-    private static final int SLEEP_TIME_WHEN_NO_TABLE_TO_COMPACT_MILLIS = 5000;
-    private static final int SLEEP_TIME_WHEN_NO_LOCK_MILLIS = 60 * 1000;
+public final class BackgroundCompactor implements AutoCloseable {
+    public static final long SLEEP_TIME_WHEN_NOTHING_TO_COMPACT_MIN_MILLIS = TimeUnit.SECONDS.toMillis(5);
 
     private static final Logger log = LoggerFactory.getLogger(BackgroundCompactor.class);
 
     private final TransactionManager transactionManager;
     private final KeyValueService keyValueService;
     private final LockService lockService;
-    private final Supplier<CompactorConfig> config;
+    private final Supplier<CompactorConfig> compactorConfigSupplier;
+    private final CompactPriorityCalculator compactPriorityCalculator;
+
+    private final CompactionOutcomeMetrics compactionOutcomeMetrics = new CompactionOutcomeMetrics();
 
     private Thread daemon;
 
-    public static void createAndRun(TransactionManager transactionManager,
+    public static Optional<BackgroundCompactor> createAndRun(TransactionManager transactionManager,
             KeyValueService keyValueService,
             LockService lockService,
             Supplier<CompactorConfig> compactorConfigSupplier) {
+        if (!keyValueService.shouldTriggerCompactions()) {
+            log.info("Not starting a background compactor, because we don't believe our KVS needs one.");
+            return Optional.empty();
+        }
+
+        CompactPriorityCalculator compactPriorityCalculator = CompactPriorityCalculator.create(transactionManager);
         BackgroundCompactor backgroundCompactor = new BackgroundCompactor(transactionManager,
                 keyValueService,
                 lockService,
-                compactorConfigSupplier);
+                compactorConfigSupplier,
+                compactPriorityCalculator);
         backgroundCompactor.runInBackground();
+
+        log.info("Created and started the background compactor.");
+
+        return Optional.of(backgroundCompactor);
     }
 
-    public BackgroundCompactor(TransactionManager transactionManager,
+    @VisibleForTesting
+    BackgroundCompactor(TransactionManager transactionManager,
             KeyValueService keyValueService,
             LockService lockService,
-            Supplier<CompactorConfig> config) {
+            Supplier<CompactorConfig> compactorConfigSupplier,
+            CompactPriorityCalculator compactPriorityCalculator) {
         this.transactionManager = transactionManager;
         this.keyValueService = keyValueService;
         this.lockService = lockService;
-        this.config = config;
-    }
-
-    public synchronized void runInBackground() {
-        Preconditions.checkState(daemon == null);
-        daemon = new Thread(this);
-        daemon.setDaemon(true);
-        daemon.setName("BackgroundCompactor");
-        daemon.start();
+        this.compactorConfigSupplier = compactorConfigSupplier;
+        this.compactPriorityCalculator = compactPriorityCalculator;
     }
 
     @Override
+    public void close() {
+        compactionOutcomeMetrics.registerOccurrenceOf(CompactionOutcome.SHUTDOWN);
+        log.info("Closing BackgroundCompactor");
+        daemon.interrupt();
+        try {
+            daemon.join();
+            daemon = null;
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw Throwables.rewrapAndThrowUncheckedException(e);
+        }
+    }
+
+    private synchronized void runInBackground() {
+        Preconditions.checkState(daemon == null);
+        daemon = new Thread(this::run);
+        daemon.setDaemon(true);
+        daemon.setName("BackgroundCompactor");
+        daemon.start();
+        log.info("Set up the background compactor to be run.");
+    }
+
     public void run() {
+        log.info("Attempting to start the background compactor for the very first time.");
         try (SingleLockService compactorLock = createSimpleLocks()) {
+            waitUntilTransactionManagerIsReady();
             log.info("Starting background compactor");
             while (true) {
                 if (Thread.currentThread().isInterrupted()) {
@@ -93,37 +121,96 @@ public class BackgroundCompactor implements Runnable {
                     return;
                 }
 
-                compactorLock.lockOrRefresh();
-                if (compactorLock.haveLocks()) {
-                    Optional<String> tableToCompactOptional = transactionManager.runTaskReadOnly(
-                            this::selectTableToCompact);
-                    if (!tableToCompactOptional.isPresent()) {
-                        log.info("No table to compact");
-                        Thread.sleep(SLEEP_TIME_WHEN_NO_TABLE_TO_COMPACT_MILLIS);
-                        continue;
-                    }
-
-                    String tableToCompact = tableToCompactOptional.get();
-
-                    log.info("Compacting table {}", LoggingArgs.safeInternalTableName(tableToCompact));
-                    compactTable(tableToCompact);
-                    log.info("Compacted table {}", LoggingArgs.safeInternalTableName(tableToCompact));
-
-                    registerCompactedTable(tableToCompact);
-                } else {
-                    log.info("Failed to get the compaction lock. Probably, another host is running compaction.");
-                    Thread.sleep(SLEEP_TIME_WHEN_NO_LOCK_MILLIS);
-                    continue;
-                }
+                runOnceRecordingOutcome(compactorLock);
             }
         } catch (InterruptedException e) {
-            log.warn("Shutting down background compactor due to InterruptedException", e);
+            log.warn("Shutting down background compactor due to InterruptedException. "
+                    + "Please restart the service to resume compactions", e);
+            compactionOutcomeMetrics.registerOccurrenceOf(CompactionOutcome.SHUTDOWN);
             Thread.currentThread().interrupt();
         }
     }
 
+    private void waitUntilTransactionManagerIsReady() throws InterruptedException {
+        while (!transactionManager.isInitialized()) {
+            log.info("Waiting for transaction manager to be initialized; going to sleep for {} ms while waiting",
+                    SafeArg.of("sleepTimeMillis", SLEEP_TIME_WHEN_NOTHING_TO_COMPACT_MIN_MILLIS));
+            Thread.sleep(SLEEP_TIME_WHEN_NOTHING_TO_COMPACT_MIN_MILLIS);
+        }
+    }
+
+    private void runOnceRecordingOutcome(SingleLockService compactorLock) throws InterruptedException {
+        CompactionOutcome outcome = CompactionOutcome.FAILED_TO_COMPACT; // default is failed, unless we know otherwise
+        try {
+            outcome = grabLockAndRunOnce(compactorLock);
+        } catch (InterruptedException e) {
+            throw e;
+        } catch (Exception e) {
+            log.warn("Unexpected exception occurred whilst performing background compaction!", e);
+        }
+        compactionOutcomeMetrics.registerOccurrenceOf(outcome);
+        Thread.sleep(getSleepTime(compactorConfigSupplier, outcome));
+    }
+
+    @VisibleForTesting
+    CompactionOutcome grabLockAndRunOnce(SingleLockService compactorLock)
+            throws InterruptedException {
+        CompactorConfig config = compactorConfigSupplier.get();
+        if (!config.enableCompaction()) {
+            log.info("Skipping compaction because it is currently disabled.");
+            return CompactionOutcome.DISABLED;
+        }
+
+        try {
+            compactorLock.lockOrRefresh();
+        } catch (Exception e) {
+            log.warn("Encountered exception when attempting to acquire the compaction lock.", e);
+            return CompactionOutcome.UNABLE_TO_ACQUIRE_LOCKS;
+        }
+        if (!compactorLock.haveLocks()) {
+            log.info("Failed to get the compaction lock. Probably, another host is running compaction.");
+            return CompactionOutcome.UNABLE_TO_ACQUIRE_LOCKS;
+        }
+
+        Optional<String> tableToCompactOptional;
+        try {
+            tableToCompactOptional = compactPriorityCalculator.selectTableToCompact();
+        } catch (Exception e) {
+            log.warn("Encountered exception when attempting to determine which table should be compacted.", e);
+            return CompactionOutcome.NOTHING_TO_COMPACT;
+        }
+        if (!tableToCompactOptional.isPresent()) {
+            log.info("No table to compact.");
+            return CompactionOutcome.NOTHING_TO_COMPACT;
+        }
+
+        String tableToCompact = tableToCompactOptional.get();
+
+        try {
+            log.info("Compacting table {}", LoggingArgs.safeInternalTableName(tableToCompact));
+            compactTable(tableToCompact, config);
+            log.info("Compacted table {}", LoggingArgs.safeInternalTableName(tableToCompact));
+        } catch (Exception e) {
+            log.warn("Encountered exception when compacting table {}",
+                    LoggingArgs.safeInternalTableName(tableToCompact),
+                    e);
+            return CompactionOutcome.FAILED_TO_COMPACT;
+        }
+
+        try {
+            registerCompactedTable(tableToCompact);
+            return CompactionOutcome.SUCCESS;
+        } catch (Exception e) {
+            log.info("Successfully compacted table {}, but failed to register this."
+                    + "Nothing bad will happen; we'll probably do a no-op compaction of this very shortly.",
+                    LoggingArgs.safeInternalTableName(tableToCompact),
+                    e);
+            return CompactionOutcome.COMPACTED_BUT_NOT_REGISTERED;
+        }
+    }
+
     private SingleLockService createSimpleLocks() {
-        return new SingleLockService(lockService, "atlas compact");
+        return SingleLockService.createSingleLockServiceWithSafeLockId(lockService, "atlas compact");
     }
 
     private void registerCompactedTable(String tableToCompact) {
@@ -137,54 +224,40 @@ public class BackgroundCompactor implements Runnable {
         });
     }
 
-    private void compactTable(String tableToCompact) {
-        keyValueService.compactInternally(TableReference.createFromFullyQualifiedName(tableToCompact));
+    private void compactTable(String tableToCompact, CompactorConfig config) {
+        // System tables MAY be involved in this process.
+        keyValueService.compactInternally(TableReference.createUnsafe(tableToCompact),
+                config.inMaintenanceMode());
     }
 
-    private Optional<String> selectTableToCompact(Transaction tx) {
-        Map<String, Long> tableToLastTimeSwept = new HashMap<>();
-        SweepPriorityTable sweepPriorityTable = SweepTableFactory.of().getSweepPriorityTable(tx);
-        sweepPriorityTable.getAllRowsUnordered(SweepPriorityTable.getColumnSelection(
-                SweepPriorityTable.SweepPriorityNamedColumn.LAST_SWEEP_TIME))
-                .forEach(row -> {
-                    Long lastSweepTime = row.getLastSweepTime();
-                    String tableName = row.getRowName().getFullTableName();
-                    tableToLastTimeSwept.put(tableName, lastSweepTime);
-                });
-
-        Map<String, Long> tableToLastTimeCompacted = new HashMap<>();
-        CompactMetadataTable compactMetadataTable = CompactTableFactory.of().getCompactMetadataTable(tx);
-        compactMetadataTable.getAllRowsUnordered(SweepPriorityTable.getColumnSelection(
-                SweepPriorityTable.SweepPriorityNamedColumn.LAST_SWEEP_TIME))
-                .forEach(row -> {
-                    Long lastCompactTime = row.getLastCompactTime();
-                    String tableName = row.getRowName().getFullTableName();
-                    tableToLastTimeCompacted.put(tableName, lastCompactTime);
-                });
-
-        List<String> uncompactedTables = tableToLastTimeSwept.keySet().stream()
-                .filter(table -> !tableToLastTimeCompacted.keySet().contains(table))
-                .collect(Collectors.toList());
-
-        if (uncompactedTables.size() > 0) {
-            int randomTableIndex = ThreadLocalRandom.current().nextInt(uncompactedTables.size());
-            return Optional.of(uncompactedTables.get(randomTableIndex));
+    @VisibleForTesting
+    static long getSleepTime(
+            Supplier<CompactorConfig> compactorConfigSupplier,
+            CompactionOutcome outcome) {
+        switch (outcome) {
+            case SUCCESS:
+            case COMPACTED_BUT_NOT_REGISTERED:
+            case SHUTDOWN:
+                return compactorConfigSupplier.get().compactPauseMillis();
+            case NOTHING_TO_COMPACT:
+            case DISABLED:
+                return Math.max(compactorConfigSupplier.get().compactPauseMillis(),
+                        SLEEP_TIME_WHEN_NOTHING_TO_COMPACT_MIN_MILLIS);
+            case UNABLE_TO_ACQUIRE_LOCKS:
+            case FAILED_TO_COMPACT:
+                return compactorConfigSupplier.get().compactPauseOnFailureMillis();
+            default:
+                throw new IllegalStateException("Unexpected outcome enum type: " + outcome);
         }
+    }
 
-        String tableToCompact = null;
-        long maxSweptAfterCompact = Long.MIN_VALUE;
-        for (Map.Entry<String, Long> entry : tableToLastTimeSwept.entrySet()) {
-            String table = entry.getKey();
-            long lastSweptTime = entry.getValue();
-            long lastCompactTime = tableToLastTimeCompacted.get(table);
-            long sweptAfterCompact = lastSweptTime - lastCompactTime;
-
-            if (sweptAfterCompact > maxSweptAfterCompact) {
-                tableToCompact = table;
-                maxSweptAfterCompact = sweptAfterCompact;
-            }
-        }
-
-        return Optional.ofNullable(tableToCompact);
+    enum CompactionOutcome {
+        SUCCESS,
+        NOTHING_TO_COMPACT,
+        COMPACTED_BUT_NOT_REGISTERED,
+        UNABLE_TO_ACQUIRE_LOCKS,
+        FAILED_TO_COMPACT,
+        SHUTDOWN,
+        DISABLED
     }
 }